metrics: rename pids_v2 to pids
dicussed in #3726
