+++ metrics/cgroups/cgroups.go
@@ -19,20 +19,13 @@
 package cgroups
 
 import (
-	"context"
-
 	"github.com/containerd/cgroups"
-	eventstypes "github.com/containerd/containerd/api/events"
-	"github.com/containerd/containerd/errdefs"
-	"github.com/containerd/containerd/events"
-	"github.com/containerd/containerd/log"
-	"github.com/containerd/containerd/namespaces"
+	v1 "github.com/containerd/containerd/metrics/cgroups/v1"
+	v2 "github.com/containerd/containerd/metrics/cgroups/v2"
 	"github.com/containerd/containerd/platforms"
 	"github.com/containerd/containerd/plugin"
 	"github.com/containerd/containerd/runtime"
-	"github.com/containerd/containerd/runtime/v1/linux"
 	metrics "github.com/docker/go-metrics"
-	"github.com/sirupsen/logrus"
 )
 
 // Config for the cgroups monitor
@@ -56,8 +49,15 @@ func New(ic *plugin.InitContext) (interface{}, error) {
 	if !config.NoPrometheus {
 		ns = metrics.NewNamespace("container", "", nil)
 	}
-	collector := newCollector(ns)
-	oom, err := newOOMCollector(ns)
+	var (
+		tm  runtime.TaskMonitor
+		err error
+	)
+	if cgroups.Mode() == cgroups.Unified {
+		tm, err = v2.NewTaskMonitor(ic.Context, ic.Events, ns)
+	} else {
+		tm, err = v1.NewTaskMonitor(ic.Context, ic.Events, ns)
+	}
 	if err != nil {
 		return nil, err
 	}
@@ -65,54 +65,5 @@ func New(ic *plugin.InitContext) (interface{}, error) {
 		metrics.Register(ns)
 	}
 	ic.Meta.Platforms = append(ic.Meta.Platforms, platforms.DefaultSpec())
-	return &cgroupsMonitor{
-		collector: collector,
-		oom:       oom,
-		context:   ic.Context,
-		publisher: ic.Events,
-	}, nil
-}
-
-type cgroupsMonitor struct {
-	collector *collector
-	oom       *oomCollector
-	context   context.Context
-	publisher events.Publisher
-}
-
-func (m *cgroupsMonitor) Monitor(c runtime.Task) error {
-	if err := m.collector.Add(c); err != nil {
-		return err
-	}
-	t, ok := c.(*linux.Task)
-	if !ok {
-		return nil
-	}
-	cg, err := t.Cgroup()
-	if err != nil {
-		if errdefs.IsNotFound(err) {
-			return nil
-		}
-		return err
-	}
-	err = m.oom.Add(c.ID(), c.Namespace(), cg, m.trigger)
-	if err == cgroups.ErrMemoryNotSupported {
-		logrus.WithError(err).Warn("OOM monitoring failed")
-		return nil
-	}
-	return err
-}
-
-func (m *cgroupsMonitor) Stop(c runtime.Task) error {
-	m.collector.Remove(c)
-	return nil
-}
-
-func (m *cgroupsMonitor) trigger(id, namespace string, cg cgroups.Cgroup) {
-	ctx := namespaces.WithNamespace(m.context, namespace)
-	if err := m.publisher.Publish(ctx, runtime.TaskOOMEventTopic, &eventstypes.TaskOOM{
-		ContainerID: id,
-	}); err != nil {
-		log.G(m.context).WithError(err).Error("post OOM event")
-	}
+	return tm, nil
 }
